// src/app/modules/configuration/finance/models/fee-type-tranches-data-view.ts
/**
 * Fee
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { SchoolFeeTypeTranchesView } from './school-fee-type-tranches-view';

/**
 * 
 */
export interface FeeTypeTranchesDataView { 
    /**
     * Gets or Sets ClassId
     */
    classId?: number;
    /**
     * Gets or Sets BoardId
     */
    schoolBoardId?: number;
    /**
     * Gets or Sets FeeTypeId
     */
    feeTypeId?: number;
    /**
     * Gets or Sets Name
     */
    feeTypeName?: string;
    /**
     * 
     */
    feeTypeTranches?: Array<SchoolFeeTypeTranchesView>;
}